Add files via upload
train_listings is the data used to make a model with the real price is listed as a variable. The test_listings data is similar data without an actual price. So I would use my model from train_listings to predict prices in test_listings. my_price_predictions is the test_listings with my predicted price for each property.